payalgo: New file for pay command.
