+++ lightningd/payalgo.h
@@ -1,0 +1,5 @@
+#ifndef LIGHTNING_LIGHTNINGD_PAYALGO_H
+#define LIGHTNING_LIGHTNINGD_PAYALGO_H
+#include "config.h"
+
+#endif /* LIGHTNING_LIGHTNINGD_PAYALGO_H */
